pyproject.toml
[tool.poetry]
name = "simples-crud-flask"
version = "0.1.0"
description = ""
authors = ["Joao Pedro Watson <joaopedrowatson@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Werkzeug = "^2.1.2"
PyJWT = "^2.4.0"
pymongo = "^4.1.1"
python-dotenv = "^0.20.0"
dnspython = "^2.2.1"

[tool.poetry.dev-dependencies]
blue = "^0.9.0"
isort = "^5.10.1"
prospector = "^1.7.7"
autopep8 = "^1.6.0"
httpie = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
